+++ server/impl/src/main/java/com/walmartlabs/concord/server/console/ConsoleService.java
@@ -20,8 +20,17 @@ package com.walmartlabs.concord.server.console;
  * =====
  */
 
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonRawValue;
 import com.walmartlabs.concord.common.validation.ConcordKey;
+import com.walmartlabs.concord.db.AbstractDao;
+import com.walmartlabs.concord.sdk.EventType;
 import com.walmartlabs.concord.server.ConcordApplicationException;
+import com.walmartlabs.concord.server.jooq.tables.ProcessCheckpoints;
+import com.walmartlabs.concord.server.jooq.tables.ProcessEvents;
+import com.walmartlabs.concord.server.jooq.tables.VProcessQueue;
 import com.walmartlabs.concord.server.metrics.WithTimer;
 import com.walmartlabs.concord.server.org.OrganizationEntry;
 import com.walmartlabs.concord.server.org.OrganizationManager;
@@ -30,6 +39,7 @@ import com.walmartlabs.concord.server.org.project.RepositoryDao;
 import com.walmartlabs.concord.server.org.secret.PasswordChecker;
 import com.walmartlabs.concord.server.org.secret.SecretDao;
 import com.walmartlabs.concord.server.org.team.TeamDao;
+import com.walmartlabs.concord.server.process.ProcessStatus;
 import com.walmartlabs.concord.server.repository.InvalidRepositoryPathException;
 import com.walmartlabs.concord.server.repository.RepositoryManager;
 import com.walmartlabs.concord.server.security.UserPrincipal;
@@ -41,6 +51,8 @@ import com.walmartlabs.concord.server.user.UserManager;
 import org.apache.shiro.SecurityUtils;
 import org.apache.shiro.authz.UnauthorizedException;
 import org.apache.shiro.subject.Subject;
+import org.jooq.*;
+import org.jooq.impl.DSL;
 import org.sonatype.siesta.Resource;
 import org.sonatype.siesta.Validate;
 
@@ -54,7 +66,15 @@ import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
+import java.sql.Timestamp;
 import java.util.*;
+
+import static com.walmartlabs.concord.server.jooq.tables.ProcessCheckpoints.PROCESS_CHECKPOINTS;
+import static com.walmartlabs.concord.server.jooq.tables.ProcessEvents.PROCESS_EVENTS;
+import static com.walmartlabs.concord.server.jooq.tables.VProcessQueue.V_PROCESS_QUEUE;
+import static org.jooq.impl.DSL.field;
+import static org.jooq.impl.DSL.function;
+import static org.jooq.impl.DSL.inline;
 
 @Named
 @Singleton
@@ -69,6 +89,7 @@ public class ConsoleService implements Resource {
     private final RepositoryDao repositoryDao;
     private final TeamDao teamDao;
     private final LdapManager ldapManager;
+    private final ProcessDao processDao;
     private final ApiKeyDao apiKeyDao;
 
     @Inject
@@ -80,6 +101,7 @@ public class ConsoleService implements Resource {
                           RepositoryDao repositoryDao,
                           TeamDao teamDao,
                           LdapManager ldapManager,
+                          ProcessDao processDao,
                           ApiKeyDao apiKeyDao) {
 
         this.projectDao = projectDao;
@@ -91,6 +113,7 @@ public class ConsoleService implements Resource {
         this.teamDao = teamDao;
         this.ldapManager = ldapManager;
         this.apiKeyDao = apiKeyDao;
+        this.processDao = processDao;
     }
 
     @GET
@@ -282,5 +305,185 @@ public class ConsoleService implements Resource {
         }
 
         return true;
+    }
+
+    @GET
+    @Path("/process")
+    @Produces(MediaType.APPLICATION_JSON)
+    @WithTimer
+    public List<ProcessEntry> listProcesses(
+            @QueryParam("orgId") UUID orgId,
+            @QueryParam("projectId") UUID projectId,
+            @QueryParam("limit") @DefaultValue("30") int limit) {
+        if (limit <= 0) {
+            throw new ConcordApplicationException("'limit' must be a positive number", Status.BAD_REQUEST);
+        }
+
+        return processDao.list(orgId, projectId, limit);
+    }
+
+    @Named
+    public static class ProcessDao extends AbstractDao {
+
+        @Inject
+        protected ProcessDao(@Named("app") Configuration cfg) {
+            super(cfg);
+        }
+
+        public List<ProcessEntry> list(UUID orgId, UUID projectId, int limit) {
+            VProcessQueue pq = V_PROCESS_QUEUE.as("pq");
+            ProcessCheckpoints pc = PROCESS_CHECKPOINTS.as("pc");
+            ProcessEvents pe = PROCESS_EVENTS.as("pe");
+
+            try (DSLContext tx = DSL.using(cfg)) {
+                Field<Object> checkpoints = tx.select(
+                        function("array_to_json", Object.class,
+                            function("array_agg", Object.class,
+                                function("json_strip_nulls", Object.class,
+                                        function("json_build_object", Object.class,
+                                                inline("id"), pc.CHECKPOINT_ID,
+                                                inline("name"), pc.CHECKPOINT_NAME,
+                                                inline("createdAt"), pc.CHECKPOINT_DATE)))))
+                        .from(pc)
+                        .where(pc.INSTANCE_ID.eq(pq.INSTANCE_ID)).asField();
+
+                Field<Object> status = tx.select(
+                        function("array_to_json", Object.class,
+                                function("array_agg", Object.class,
+                                        function("json_strip_nulls", Object.class,
+                                                function("json_build_object", Object.class,
+                                                        inline("changeDate"), pe.EVENT_DATE,
+                                                        inline("status"), field("{0}->'status'", Object.class, pe.EVENT_DATA),
+                                                        inline("checkpointId"), field("{0}->'checkpointId'", Object.class, pe.EVENT_DATA))))))
+                        .from(pe)
+                        .where(pq.INSTANCE_ID.eq(pe.INSTANCE_ID).and(pe.EVENT_TYPE.eq(EventType.PROCESS_STATUS.name()).and(pe.EVENT_DATE.greaterOrEqual(pq.CREATED_AT))))
+                        .asField();
+
+                SelectJoinStep<Record11<UUID, UUID, String, UUID, String, String, String, Timestamp, Timestamp, Object, Object>> s = tx
+                        .select(pq.INSTANCE_ID,
+                                pq.ORG_ID, pq.ORG_NAME,
+                                pq.PROJECT_ID, pq.PROJECT_NAME,
+                                pq.INITIATOR,
+                                pq.CURRENT_STATUS,
+                                pq.CREATED_AT,
+                                pq.LAST_UPDATED_AT,
+                                checkpoints,
+                                status)
+                        .from(pq);
+
+                if (orgId != null) {
+                    s.where(pq.ORG_ID.eq(orgId));
+                }
+
+                if (projectId != null) {
+                    s.where(pq.PROJECT_ID.eq(projectId));
+                }
+
+                return s.orderBy(pq.CREATED_AT.desc())
+                        .limit(limit)
+                        .fetch(this::toEntry);
+            }
+        }
+
+        private ProcessEntry toEntry(Record11<UUID, UUID, String, UUID, String, String, String, Timestamp, Timestamp, Object, Object> r) {
+            return new ProcessEntry(r.value1(),
+                    r.value2(), r.value3(),
+                    r.value4(), r.value5(),
+                    r.value6(),
+                    ProcessStatus.valueOf(r.value7()),
+                    r.value8(),
+                    r.value9(),
+                    r.value10(),
+                    r.value11());
+        }
+    }
+
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    private static class ProcessEntry {
+        private final UUID instanceId;
+        private final UUID orgId;
+        private final String orgName;
+        private final UUID projectId;
+        private final String projectName;
+        private final String initiator;
+        private final ProcessStatus status;
+
+        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSX")
+        private final Date createdAt;
+        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSX")
+        private final Date lastUpdatedAt;
+
+        private final Object checkpoints;
+        private final Object statusHistory;
+
+        public ProcessEntry(@JsonProperty("instanceId") UUID instanceId,
+                            @JsonProperty("orgId") UUID orgId,
+                            @JsonProperty("orgName") String orgName,
+                            @JsonProperty("projectId") UUID projectId,
+                            @JsonProperty("projectName") String projectName,
+                            @JsonProperty("initiator") String initiator,
+                            @JsonProperty("status") ProcessStatus status,
+                            @JsonProperty("createdAt") Date createdAt,
+                            @JsonProperty("lastUpdatedAt") Date lastUpdatedAt,
+                            @JsonProperty("checkpoints") Object checkpoints,
+                            @JsonProperty("history") Object statusHistory) {
+            this.instanceId = instanceId;
+            this.orgId = orgId;
+            this.orgName = orgName;
+            this.projectId = projectId;
+            this.projectName = projectName;
+            this.initiator = initiator;
+            this.status = status;
+            this.createdAt = createdAt;
+            this.lastUpdatedAt = lastUpdatedAt;
+            this.checkpoints = checkpoints;
+            this.statusHistory = statusHistory;
+        }
+
+        public UUID getInstanceId() {
+            return instanceId;
+        }
+
+        public UUID getOrgId() {
+            return orgId;
+        }
+
+        public String getOrgName() {
+            return orgName;
+        }
+
+        public UUID getProjectId() {
+            return projectId;
+        }
+
+        public String getProjectName() {
+            return projectName;
+        }
+
+        public String getInitiator() {
+            return initiator;
+        }
+
+        public ProcessStatus getStatus() {
+            return status;
+        }
+
+        public Date getCreatedAt() {
+            return createdAt;
+        }
+
+        public Date getLastUpdatedAt() {
+            return lastUpdatedAt;
+        }
+
+        @JsonRawValue
+        public Object getCheckpoints() {
+            return checkpoints;
+        }
+
+        @JsonRawValue
+        public Object getStatusHistory() {
+            return statusHistory;
+        }
     }
 }